+++ src/pages/administration/students/index.vue
@@ -13,7 +13,7 @@
 		/>
 		<backend-data-table
 			:actions="permissionFilteredTableActions"
-			:columns="tableColumns"
+			:columns="filteredColumns"
 			:current-page.sync="page"
 			:data="students"
 			:paginated="true"
@@ -60,10 +60,7 @@
 					/>
 				</span>
 			</template>
-			<template
-				v-if="schoolInternallyManaged"
-				v-slot:datacolumn-_id="{ data, selected, highlighted }"
-			>
+			<template v-slot:datacolumn-_id="{ data, selected, highlighted }">
 				<base-button
 					:class="{
 						'action-button': true,
@@ -283,6 +280,12 @@ export default {
 			return this.tableActions.filter((action) =>
 				action.permission ? this.$_userHasPermission(action.permission) : true
 			);
+		},
+		filteredColumns() {
+			// filters edit column if school is external
+			return this.school.isExternal
+				? this.tableColumns.filter((col) => col.field !== "_id")
+				: this.tableColumns;
 		},
 	},
 	watch: {